extract IngredientPouchModel from InventoryModel
