+++ common/src/main/java/com/wynntils/models/inventory/InventoryModel.java
@@ -11,12 +11,10 @@ import com.wynntils.mc.event.ContainerSetContentEvent;
 import com.wynntils.mc.event.SetSlotEvent;
 import com.wynntils.models.inventory.type.InventoryAccessory;
 import com.wynntils.models.items.items.game.GearItem;
-import com.wynntils.models.items.items.gui.IngredientPouchItem;
 import com.wynntils.models.worlds.event.WorldStateEvent;
 import com.wynntils.models.worlds.type.WorldState;
 import com.wynntils.utils.mc.McUtils;
 import com.wynntils.utils.type.CappedValue;
-import com.wynntils.utils.wynn.InventoryUtils;
 import com.wynntils.utils.wynn.ItemUtils;
 import java.util.ArrayList;
 import java.util.Collections;
@@ -30,7 +28,6 @@ import net.neoforged.bus.api.SubscribeEvent;
 
 public final class InventoryModel extends Model {
     private static final int MAX_INVENTORY_SLOTS = 28;
-    private static final int MAX_INGREDIENT_POUCH_SLOTS = 27;
     private static final List<String> AUTO_CASTER_MAJOR_IDS = List.of("Sorcery", "Madness");
 
     private final InventoryWatcher emptySlotWatcher = new InventoryWatcher(ItemStack::isEmpty);
@@ -42,10 +39,6 @@ public final class InventoryModel extends Model {
 
     public CappedValue getInventorySlots() {
         return new CappedValue(MAX_INVENTORY_SLOTS - getEmptySlots(), MAX_INVENTORY_SLOTS);
-    }
-
-    public CappedValue getIngredientPouchSlots() {
-        return new CappedValue(getUsedIngredientPouchSlots(), MAX_INGREDIENT_POUCH_SLOTS);
     }
 
     public void registerWatcher(InventoryWatcher watcher) {
@@ -151,16 +144,6 @@ public final class InventoryModel extends Model {
 
     private int getEmptySlots() {
         return emptySlotWatcher.getSlots();
-    }
-
-    private int getUsedIngredientPouchSlots() {
-        ItemStack itemStack = McUtils.inventory().items.get(InventoryUtils.INGREDIENT_POUCH_SLOT_NUM);
-        Optional<IngredientPouchItem> pouchItemOpt = Models.Item.asWynnItem(itemStack, IngredientPouchItem.class);
-
-        // This should never happen
-        if (pouchItemOpt.isEmpty()) return -1;
-
-        return pouchItemOpt.get().getIngredients().size();
     }
 
     private void updateCache() {
